Rewrite CSV loading process
SmarterCSV store each CSV row as its own hash, but it turns out that it doesn't
work properly when there is no default headers (such as with our CSV file).
Thus, it may be better to replace the SmarterCSV gem with the CSV standard libary,
and then write some code to generate the hashes with the proper metadata that I
plan to use for the heuristics.